upgrade to flow-bin 0.95.1 and fix indent-string libdef
This upgrades to flow-bin 0.95.1, which includes an update to `JSON.stringify`.
This update means `stringify` can return `string | undefined` (`undefined` is
covered by a few edge cases). Flow does not accept `undefined` in string
interpolations, so the string "undefined" has been used in these cases as a fix.

`indent-string` changed its signature in 3.2.0. The updated libdef has been
submitted to `flow-typed` and here we pull it down. Uses of the `indent` call
have been fixed.  It is suspected that the upgrade of Flow has increased type
coverage here and flagged this as an error, because this should have otherwise
gone undetected.